Fix splice1 and splice2 param len type: Use unsigned int

+++ src/cat.c
@@ -39,7 +39,7 @@ static bool io_uring_has_cqe(struct io_uring *ring)
 /**
  * in_fd or out_fd is a pipe, so use spice directly
  */
-static int splice1(struct io_uring *ring, int in_fd, size_t len)
+static int splice1(struct io_uring *ring, int in_fd, unsigned len)
 {
     for (struct io_uring_sqe *sqe; (sqe = io_uring_get_sqe(ring)); ) {
         io_uring_prep_splice(sqe, in_fd, (loff_t) -1, out_fd, (loff_t) -1, len, 0);
@@ -68,7 +68,7 @@ static int splice1(struct io_uring *ring, int in_fd, size_t len)
 /**
  * Neither in_fd nor out_fd is a pipe, so create one and spice.
  */
-int splice2(struct io_uring *ring, int in_fd, size_t len)
+int splice2(struct io_uring *ring, int in_fd, unsigned len)
 {
     int pipefd[2];
     if (pipe(pipefd) < 0) {
